@@ -9,21 +9,21 @@ homepage: https://github.com/ayalma/feature_discovery
 environment:
   # This should match the Flutter stable SDK constraint.
   # See https://github.com/flutter/flutter/blob/stable/packages/flutter/pubspec.yaml.
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.0.4
-  shared_preferences: ^0.5.6+1
+  provider: ^5.0.0
+  shared_preferences: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  pedantic: 1.9.0
+  pedantic: 1.11.0
 
 dependency_overrides:
   # TODO(creativecreatorormaybenot): Remove when Pedantic is updated in flutter_test.
-  pedantic: 1.9.0
+  pedantic: 1.11.0
