@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
 /// We use [RenderProxyBox] because we only want to clip and keep
@@ -8,11 +7,9 @@ class RenderClipContent extends RenderProxyBox {
   double _radius;
 
   RenderClipContent({
-    @required Offset center,
-    @required double radius,
-  })  : assert(center != null),
-        assert(radius != null),
-        _center = center,
+    required Offset center,
+    required double radius,
+  })   : _center = center,
         _radius = radius;
 
   /// The inner area of the DescribedFeatureOverlay.
@@ -37,11 +34,11 @@ class RenderClipContent extends RenderProxyBox {
   /// The reason this is necessary is that the content that might be overflowing will catch
   /// the hit events even when it is clipped out in paint.
   @override
-  bool hitTest(BoxHitTestResult result, {Offset position}) {
+  bool hitTest(BoxHitTestResult result, {Offset? position}) {
     // If the hit is inside of the inner area of the DescribedFeatureOverlay,
     // we want to catch the hit event and pass it to the children. Otherwise, we want to ignore it in order
     // to allow the GestureDetector in DescribedFeatureOverlay to catch it.
-    if (innerCircle.contains(position) &&
+    if (innerCircle.contains(position!) &&
         hitTestChildren(result, position: position)) {
       result.add(BoxHitTestEntry(this, position));
       return true;
