@@ -4,7 +4,7 @@ class CenterAbout extends StatelessWidget {
   final Offset position;
   final Widget child;
 
-  const CenterAbout({Key key, this.position, this.child}) : super(key: key);
+  const CenterAbout({Key? key, required this.position, required this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => Positioned(
@@ -23,7 +23,7 @@ class AnchoredOverlay extends StatelessWidget {
   final Widget child;
 
   const AnchoredOverlay(
-      {Key key, this.showOverlay, this.overlayBuilder, this.child})
+      {Key? key, required this.showOverlay, required this.overlayBuilder, required this.child,})
       : super(key: key);
 
   @override
@@ -43,13 +43,15 @@ class AnchoredOverlay extends StatelessWidget {
       );
 }
 
+typedef OverlayBuilderBuilder = Widget Function(BuildContext context);
+
 class OverlayBuilder extends StatefulWidget {
-  final bool showOverlay;
-  final Function(BuildContext context) overlayBuilder;
+  final bool? showOverlay;
+  final OverlayBuilderBuilder overlayBuilder;
   final Widget child;
 
   const OverlayBuilder(
-      {Key key, this.showOverlay = false, this.overlayBuilder, this.child})
+      {Key? key, this.showOverlay = false, required this.overlayBuilder, required this.child,})
       : super(key: key);
 
   @override
@@ -57,12 +59,12 @@ class OverlayBuilder extends StatefulWidget {
 }
 
 class _OverlayBuilderState extends State<OverlayBuilder> {
-  OverlayEntry overlayEntry;
+  OverlayEntry? overlayEntry;
 
   @override
   void initState() {
     super.initState();
-    if (widget.showOverlay) showOverlay();
+    if (widget.showOverlay!) showOverlay();
   }
 
   @override
@@ -89,27 +91,27 @@ class _OverlayBuilderState extends State<OverlayBuilder> {
     overlayEntry = OverlayEntry(
       builder: widget.overlayBuilder,
     );
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      Overlay.of(context).insert(overlayEntry);
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
+      Overlay.of(context)!.insert(overlayEntry!);
     });
   }
 
   void hideOverlay() {
-    overlayEntry.remove();
+    overlayEntry!.remove();
     overlayEntry = null;
   }
 
   void syncWidgetAndOverlay() {
-    if (isShowingOverlay() && !widget.showOverlay) {
+    if (isShowingOverlay() && !widget.showOverlay!) {
       hideOverlay();
-    } else if (!isShowingOverlay() && widget.showOverlay) showOverlay();
+    } else if (!isShowingOverlay() && widget.showOverlay!) showOverlay();
   }
 
   void buildOverlay() async => overlayEntry?.markNeedsBuild();
 
   @override
   Widget build(BuildContext context) {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       buildOverlay();
     });
     return widget.child;
